New script to run prior to splitobs_part1, etc
In a subset of pipeline calibrated datasets (CASA 4.7.*), the standard ms.split.cal end product are *not* created, and instead [files].ms are created which house the calibrated information in a special column.  In these circumstances, splitobs_part1.py will not work, because there are no ms.split.cal files to work with.  This new script should be run prior to splitobs_part1.py, in the same version of CASA as was used for the original calibration.  In the special cases described above (only), it will create the standard ms.split.cal files and allow splitobs_part1.py to work normally afterward.